// src/shared/infrastructure/persistence/generated.ts
export type Decimal = ColumnType<string, number | string>;

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Json = ColumnType<JsonValue, string, string>;

export type JsonArray = JsonValue[];

export type JsonObject = {
  [x: string]: JsonValue | undefined;
};

export type JsonPrimitive = boolean | number | string | null;

export type JsonValue = JsonArray | JsonObject | JsonPrimitive;

export interface Accounts {
  balance: Generated<Decimal>;
  code: string;
  created_at: Date | null;
  deleted_at: Date | null;
  id: Generated<number>;
  name: string;
  notes: string | null;
  parent_id: number | null;
  status: Generated<string>;
  type_id: number | null;
  updated_at: Date | null;
}

export interface AccountTypes {
  basecode: string;
  created_at: Date | null;
  credit: number | null;
  debit: number | null;
  deleted_at: Date | null;
  id: Generated<number>;
  name: string;
  type: string | null;
  updated_at: Date | null;
}

export interface ActivityLog {
  batch_uuid: string | null;
  causer_id: number | null;
  causer_type: string | null;
  created_at: Date | null;
  description: string;
  event: string | null;
  id: Generated<number>;
  log_name: string | null;
  properties: Json | null;
  subject_id: number | null;
  subject_type: string | null;
  updated_at: Date | null;
}

export interface Cache {
  expiration: number;
  key: string;
  value: string;
}

export interface CacheLocks {
  expiration: number;
  key: string;
  owner: string;
}

export interface Companies {
  address: Json | null;
  code: string | null;
  created_at: Date | null;
  database: string | null;
  deleted_at: Date | null;
  id: Generated<number>;
  name: string;
  updated_at: Date | null;
}

export interface CompaniesUsers {
  company_id: number;
  created_at: Date | null;
  deleted_at: Date | null;
  id: Generated<number>;
  notes: string | null;
  status: Generated<string>;
  updated_at: Date | null;
  user_id: number;
}

export interface FailedJobs {
  connection: string;
  exception: string;
  failed_at: Generated<Date>;
  id: Generated<number>;
  payload: string;
  queue: string;
  uuid: string;
}

export interface Groups {
  address: Json | null;
  code: string | null;
  created_at: Date | null;
  deleted_at: Date | null;
  id: Generated<number>;
  name: string;
  notes: string | null;
  status: Generated<string>;
  updated_at: Date | null;
}

export interface Inventories {
  balance: Generated<Decimal>;
  code: string | null;
  cost_per_unit: Generated<Decimal>;
  created_at: Date | null;
  deleted_at: Date | null;
  expiry_date: Date | null;
  id: Generated<number>;
  item_id: number | null;
  item_type: string | null;
  location_id: number | null;
  location_type: string | null;
  model_id: number | null;
  model_type: string | null;
  name: string | null;
  notes: string | null;
  parent_id: number | null;
  parent_type: string | null;
  quantity: Generated<number>;
  sku: string | null;
  space_id: number | null;
  space_type: string | null;
  status: Generated<string>;
  type_id: number | null;
  type_type: string | null;
  updated_at: Date | null;
}

export interface Items {
  attributes: Json | null;
  code: string | null;
  cost: Generated<Decimal>;
  created_at: Date | null;
  deleted_at: Date | null;
  description: string | null;
  dimension: Json | null;
  files: Json | null;
  id: Generated<number>;
  images: Json | null;
  links: Json | null;
  model_id: number | null;
  model_type: string | null;
  name: string;
  notes: string | null;
  options: Json | null;
  parent_id: number | null;
  parent_type: string | null;
  price: Generated<Decimal>;
  primary_code: string | null;
  sku: string | null;
  space_id: number | null;
  space_type: string | null;
  status: Generated<string>;
  tags: Json | null;
  type_id: number | null;
  type_type: string | null;
  updated_at: Date | null;
  variants: Json | null;
  weight: Generated<Decimal>;
}

export interface JobBatches {
  cancelled_at: number | null;
  created_at: number;
  failed_job_ids: string;
  failed_jobs: number;
  finished_at: number | null;
  id: string;
  name: string;
  options: string | null;
  pending_jobs: number;
  total_jobs: number;
}

export interface Jobs {
  attempts: number;
  available_at: number;
  created_at: number;
  id: Generated<number>;
  payload: string;
  queue: string;
  reserved_at: number | null;
}

export interface Migrations {
  batch: number;
  id: Generated<number>;
  migration: string;
}

export interface ModelHasPermissions {
  model_id: number;
  model_type: string;
  permission_id: number;
  team_id: number | null;
}

export interface ModelHasRoles {
  model_id: number;
  model_type: string;
  role_id: number;
  team_id: number | null;
}

export interface PasswordResetTokens {
  created_at: Date | null;
  email: string;
  token: string;
}

export interface Permissions {
  created_at: Date | null;
  deleted_at: Date | null;
  guard_name: string;
  id: Generated<number>;
  name: string;
  updated_at: Date | null;
}

export interface PersonalAccessTokens {
  abilities: string | null;
  created_at: Date | null;
  expires_at: Date | null;
  id: Generated<number>;
  last_used_at: Date | null;
  name: string;
  token: string;
  tokenable_id: number;
  tokenable_type: string;
  updated_at: Date | null;
}

export interface Persons {
  address: Json | null;
  address_detail: string | null;
  birth_date: Date | null;
  country: string | null;
  created_at: Date | null;
  death_date: Date | null;
  deleted_at: Date | null;
  district: string | null;
  email: string | null;
  full_name: string | null;
  gender: string | null;
  id: Generated<number>;
  id_card_number: string | null;
  name: string;
  notes: string | null;
  number: string | null;
  phone_number: string | null;
  postal_code: string | null;
  province: string | null;
  regency: string | null;
  status: Generated<string>;
  sub_district: string | null;
  updated_at: Date | null;
  village: string | null;
}

export interface Players {
  address: Json | null;
  address_detail: string | null;
  code: string | null;
  country: string | null;
  created_at: Date | null;
  deleted_at: Date | null;
  district: string | null;
  email: string | null;
  files: Json | null;
  id: Generated<number>;
  links: Json | null;
  name: string;
  notes: string | null;
  phone_number: string | null;
  postal_code: string | null;
  province: string | null;
  regency: string | null;
  shopee_username: string | null;
  size_id: number | null;
  size_type: string | null;
  space_id: number | null;
  space_type: string | null;
  status: Generated<string>;
  sub_district: string | null;
  tags: Json | null;
  tokopedia_username: string | null;
  type_id: number | null;
  type_type: string | null;
  updated_at: Date | null;
  village: string | null;
  whatsapp_number: string | null;
}

export interface Relations {
  created_at: Date | null;
  deleted_at: Date | null;
  id: Generated<number>;
  model1_id: number | null;
  model1_quantity: number | null;
  model1_type: string | null;
  model2_id: number | null;
  model2_quantity: number | null;
  model2_type: string | null;
  name: string | null;
  notes: string | null;
  number: string | null;
  status: Generated<string>;
  type: string | null;
  updated_at: Date | null;
}

export interface RoleHasPermissions {
  permission_id: number;
  role_id: number;
}

export interface RolePermissions {
  created_at: Date | null;
  deleted_at: Date | null;
  id: Generated<number>;
  permission_id: number;
  updated_at: Date | null;
}

export interface Roles {
  created_at: Date | null;
  deleted_at: Date | null;
  guard_name: string;
  id: Generated<number>;
  name: string;
  team_id: number | null;
  updated_at: Date | null;
}

export interface Sessions {
  id: string;
  ip_address: string | null;
  last_activity: number;
  payload: string;
  user_agent: string | null;
  user_id: number | null;
}

export interface Shareholders {
  company_id: number;
  created_at: Date | null;
  deleted_at: Date | null;
  id: Generated<number>;
  notes: string | null;
  owner_id: number;
  owner_type: string;
  shares: Generated<string>;
  updated_at: Date | null;
}

export interface Spaces {
  address: Json | null;
  code: string | null;
  created_at: Date | null;
  deleted_at: Date | null;
  files: Json | null;
  id: Generated<number>;
  name: string;
  notes: string | null;
  parent_id: number | null;
  parent_type: string | null;
  status: Generated<string>;
  type_id: number | null;
  type_type: string | null;
  updated_at: Date | null;
}

export interface TransactionDetails {
  cost_per_unit: Generated<Decimal>;
  created_at: Date | null;
  credit: Generated<Decimal>;
  data: Json | null;
  debit: Generated<Decimal>;
  deleted_at: Date | null;
  detail_id: number | null;
  detail_type: string | null;
  discount: Generated<Decimal>;
  id: Generated<number>;
  model_id: number | null;
  model_type: string | null;
  name: string | null;
  notes: string | null;
  price: Generated<Decimal>;
  quantity: Generated<Decimal>;
  sku: string | null;
  transaction_id: number;
  updated_at: Date | null;
  volume: Generated<Decimal>;
  weight: Generated<Decimal>;
}

export interface Transactions {
  class: string | null;
  created_at: Date | null;
  deleted_at: Date | null;
  description: string | null;
  fee: Generated<Decimal>;
  fee_rules: string | null;
  files: Json | null;
  handler_id: number | null;
  handler_notes: string | null;
  handler_number: string | null;
  handler_type: string | null;
  id: Generated<number>;
  input_address: Json | null;
  input_id: number | null;
  input_type: string | null;
  links: Json | null;
  model_id: number | null;
  model_type: string | null;
  notes: string | null;
  number: string | null;
  output_address: Json | null;
  output_id: number | null;
  output_type: string | null;
  parent_id: number | null;
  parent_type: string | null;
  received_time: Date | null;
  receiver_id: number | null;
  receiver_notes: string | null;
  receiver_type: string | null;
  relation_id: number | null;
  relation_type: string | null;
  request_time: Date | null;
  sender_id: number | null;
  sender_notes: string | null;
  sender_type: string | null;
  sent_date: Date | null;
  sent_time: Date | null;
  space_id: number | null;
  space_type: string | null;
  status: Generated<string>;
  tags: Json | null;
  total: Generated<Decimal>;
  total_details: Generated<Decimal>;
  type_id: number | null;
  type_type: string | null;
  updated_at: Date | null;
}

export interface Users {
  address: Json | null;
  birth_date: Date | null;
  created_at: Date | null;
  deleted_at: Date | null;
  email: string;
  email_verified_at: Date | null;
  id: Generated<number>;
  id_card_number: string | null;
  left_date: Date | null;
  name: string;
  password: string;
  phone_number: string | null;
  player_id: number | null;
  remember_token: string | null;
  role_id: number | null;
  sex: string | null;
  status: Generated<string>;
  updated_at: Date | null;
  username: string | null;
}
// ...
export interface DB {
  account_types: AccountTypes;
  accounts: Accounts;
  activity_log: ActivityLog;
  cache: Cache;
  cache_locks: CacheLocks;
  companies: Companies;
  companies_users: CompaniesUsers;
  failed_jobs: FailedJobs;
  groups: Groups;
  inventories: Inventories;
  items: Items;
  job_batches: JobBatches;
  jobs: Jobs;
  migrations: Migrations;
  model_has_permissions: ModelHasPermissions;
  model_has_roles: ModelHasRoles;
  password_reset_tokens: PasswordResetTokens;
  permissions: Permissions;
  personal_access_tokens: PersonalAccessTokens;
  persons: Persons;
  players: Players;
  relations: Relations;
  role_has_permissions: RoleHasPermissions;
  role_permissions: RolePermissions;
  roles: Roles;
  sessions: Sessions;
  shareholders: Shareholders;
  spaces: Spaces;
  transaction_details: TransactionDetails;
  transactions: Transactions;
  users: Users;
  variables: Variables;
}